Python_Batch_Remove_Images_Background/requirements.txt
onnxruntime
rembg
pillow